+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the convention described at
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- `query_greater_than()` like `query_range()` with only the lower key.
+- `query_less_than()` like `query_range()` with only the upper key.
+
 ## [2.1.1] - 2020-04-24
 ### Changed
 - Upgrade to `rust-rocksdb` 0.14.0 release.